refactor(operators): inherit UR operator from BaseOperator
