+++ configs/pi0/pi0_paligemma_ur3_full_finetune.py
@@ -277,7 +277,7 @@ inference = dict(
         norm_type='proprio',
     ),
     operator=dict(
-        type='NewUROperator',
+        type='UROperator',
         command_mode='joint',
         img_left_topic='/wrist_camera/color/image_raw',
         img_front_topic='/front_camera/color/image_raw',
