+++ configs/pi05/pi05_paligemma_ur3_full_finetune.py
@@ -261,7 +261,7 @@ inference = dict(
     ),
     action_chunk=50,
     operator=dict(
-        type='NewUROperator',
+        type='UROperator',
         command_mode='joint',
         img_left_topic='/wrist_camera/color/image_raw',
         img_front_topic='/front_camera/color/image_raw',
